+++ ui/owlbridge.py
@@ -1120,7 +1120,8 @@ def genDic(plan):
     # Should we ignore expert options that will reset to default?
     optList = ["netSpending", "maxIter", "maxRothConversion", "maxTime", "noRothConversions",
                "startRothConversions", "withMedicare", "bequest", "solver", "noLateSurplus",
-               "spendingSlack", "oppCostX", "amoConstraints", "amoRoth", "amoSurplus", "withSCLoop",
+               "spendingSlack", "spendingWeight", "spendingFloor", "timePreference", "oppCostX",
+               "amoConstraints", "amoRoth", "amoSurplus", "withSCLoop",
                "absTol", "bigMamo", "relTol",]
     for key in optList:
         if key in solverOptionKeys:
@@ -1156,6 +1157,8 @@ def genDic(plan):
 
     if plan.objective == "maxSpending":
         dic["objective"] = "Net spending"
+    elif plan.objective == "maxHybrid":
+        dic["objective"] = "Hybrid"
     else:
         dic["objective"] = "Bequest"